Feat: 메인 페이지의 API 요청에 staleTime, gcTime 추가
