+++ src/features/main/queries/getRecommendedTracks.ts
@@ -1,6 +1,7 @@
 import { RECOMMENDED_TRACK_IDS } from "@/constants/spotify";
 import { spotifyFetch } from "@/lib/spotify-api-client";
 import { SpotifyTrack } from "@/types/spotify";
+import { useQuery } from "@tanstack/react-query";
 
 /**
  * 추천 트랙 정보를 가져오는 함수
@@ -24,4 +25,20 @@ export async function getRecommendedTracks(): Promise<SpotifyTrack[]> {
     );
     throw error;
   }
+}
+
+/**
+ * 추천 트랙 데이터를 가져오는 훅
+ * 캐싱 기능 제공으로 중복 요청 방지 및 성능 개선
+ */
+export function useRecommendedTracks() {
+  return useQuery<SpotifyTrack[], Error>({
+    queryKey: ["recommendedTracks"],
+    queryFn: getRecommendedTracks,
+    staleTime: 30 * 60 * 1000, // 30분 동안 캐시 유지
+    gcTime: 60 * 60 * 1000, // 1시간 동안 캐시 데이터 유지
+    refetchOnWindowFocus: false,
+    refetchOnMount: false,
+    retry: 1,
+  });
 }
